[FEATURE] Adds the ability to debug the wallet as another user
